Bump version number to 230
So distributions asked for us to bump the version number early, so a
libgudev package will replace the gudev sub-package provided via systemd.
This is not strictly required, as we will keep patching systemd in
parallel to libgudev, and provide compatible releases. However, there's no
reason to not fulfil this demand. So bump libgudev to 230.

We intend to drop gudev from systemd-221, but lets be safe and bump the
version to 230 now.

https://bugzilla.gnome.org/show_bug.cgi?id=750233