gudev: Add g_udev_client_query_by_subsystem_and_device_file()
Looks up a device for a device file for a particular subsystem. In contrast to
g_udev_client_query_device_file() this function does not make blocking I/O
through stat()'ing a device file, and relies solely on the udev database making
it a better fit for mocked sysfs hierarchies.