device: add g_udev_device_get_current_tags
Starting with systemd-udevd 247 tags are "sticky" meaning that once
a tag is assigned to a device it will not be removed from the device
again until the device itself is removed.

New property CURRENT_TAGS has been added that works similar to the
existing TAGS property but only lists tags set by the most recent
uevent/database update.

https://lists.freedesktop.org/archives/systemd-devel/2020-November/045646.html